refactor: Unifier le renvoi d'email dans la route email-[token]
Simplification majeure en utilisant une seule route au lieu de deux.

Avant:
- /auth/verify/email-[token] → vérification d'email
- /auth/verify/resend-email-[email] → renvoi d'email (route séparée)

Après:
- /auth/verify/email-[token] → vérification d'email (par défaut)
- /auth/verify/email-[email]?resend=true → renvoi d'email (nouveau mode)

Changements:
1. Modification de /auth/verify/email-[token]/+page.server.ts
   - Ajout de la détection du paramètre 'resend'
   - Si resend=true: le token est interprété comme un email encodé
   - Logique de renvoi d'email avec validation et redirection
   - Gestion d'erreur robuste (utilisateur/token manquant)

2. Modification de /auth/verify/email/+page.svelte
   - Lien 'clickez ici' pointe maintenant vers email-[token]?resend=true
   - Utilisation de resolve() avec le token = email encodé

3. Suppression de /auth/verify/resend-email-[email]/
   - Route complètement supprimée (dossier + fichiers)
   - Plus de duplication de code
   - Plus de structure de données incompatible

Avantages:
✅ Une seule route à maintenir au lieu de deux
✅ Code centralisé et DRY (Don't Repeat Yourself)
✅ Logique claire avec if (resend === 'true')
✅ Moins de fichiers et de complexité
✅ Gestion d'erreur cohérente avec messages flash
✅ Solution plus élégante et maintenable

Comportement:
- Sans paramètre: vérification normale avec token
- Avec ?resend=true: renvoi d'email (token = email)
- Redirection vers /auth/verify/email avec toast de succès/erreur
- Logs console pour le débogage
